+++ @internal/ssr/src/classes/fetch/request/Requests.js
@@ -1,5 +1,5 @@
+
 import { Request } from './Request';
-import { Request as NewRequest } from './NewRequest';
 
 class Requests {
   requests = [];
@@ -9,12 +9,8 @@ class Requests {
     this.getNext = this.getNext.bind(this);
   }
 
-  create(url, config, resolve, doRequest) {
-    this.requests.push(new Request(url, config, resolve, doRequest));
-  }
-
-  newCreate(url, config, ctx) {
-    this.requests.push(new NewRequest(url, config, ctx));
+  create(url, config, ctx) {
+    this.requests.push(new Request(url, config, ctx));
   }
 
   getAll() {
