--- @internal/ssr/src/classes/fetch/response/Response.js
@@ -1,15 +1,0 @@
-import { SyncPromise } from '../../base/SyncPromise';
-
-class Response {
-  constructor(getter) {
-    this.json = () => new SyncPromise(resolve => {
-      resolve(getter());
-    });
-
-    this.text = () => new SyncPromise(resolve => {
-      resolve(getter());
-    });
-  }
-}
-
-export { Response };
